alterei as structs node e BinaryTree para tree_utils
