+++ src/bst.h
@@ -6,22 +6,9 @@
 #include <vector>
 #include <string>
 #include <chrono>
+#include "tree_utils.h"
 
 namespace BST{
-    struct Node {
-        std::string word;
-        std::vector<int> documentIds;
-        Node* parent;
-        Node* left;
-        Node* right;
-        int height;    // usado na AVL
-        int isRed;     // usado na RBT
-    };
-
-    struct BinaryTree {
-        Node* root;
-        Node* NIL;  // usado na RBT (Opcional)
-    };
 
     struct InsertResult {
         int numComparisons;
